Relocate Kubernetes models
