+++ src/model/migrate.go
@@ -30,7 +30,6 @@ func Migrate(db *gorm.DB, logger logrus.FieldLogger) error {
 		&AmazonNodePoolsModel{},
 		&EKSClusterModel{},
 		&EKSSubnetModel{},
-		&KubernetesClusterModel{},
 	}
 
 	var tableNames string
